Define simple structure for the AST,
with a few different expression types.
I'm starting out with the first mechanism I've done (and faced
problems with) in the Monkey lang as well; a simple OO-style
inheritance structure with a single base class to identify all
possible variants.
Not sure whether this is going to cause problems again down the line
yet.