WIP on the visitor pattern experimentation.
The implementation of the pattern itself is not quite complex.
The problem lies with the need to have the specialized types
(BinaryExpr, Grouping, etc) to have a member of type Expr, which is
the base type of all those specialized types. And the base type is
supposed to define a pure virtual function which "accepts" the
visitor. The compiler isn't quite happy about that.