Using the std::visit alongside the std::variant,
instead of the theorized OO-pattern, in order to implement the
"visitor" pattern. Which actually turns out to be way simpler in my
opinion.
And I bet this will potentially generate more leaner binary in the
end, due to the fact that there's zero inheritance-hierarchy and
zero virtual function calls, which would potentially eliminate the
"double dispatch" shenanigan that typically comes with the "visitor"
design pattern.

Another thing to note here is the use of "unique_ptr<Expr>" to
handle the self-referential issue on the need for the specialized
expression type (BinaryExpr, for instance) to have the generalized
Expr type as a member. In addition to the ownership semantics, this
method allows to indicate the possibility of the Expr type members
being blank/null/undefined etc.