Create a simple function to test out the AST debug printing.
Looks good so far.

Found a small issue that might need to be worked out next.
The current capability won't allow composing independent expressions
to be composed separately and assign them as part of a larger
expression. (Mainly due to the fact that the internal expression
references are primarily unique pointers.) So, basically none of the
expressions are copiable/assignable to another expression tree.
Might wanna work on the copy semantics next.