Refactoring out the current AST structure into something simple,
which is essentially C-unions and raw pointers. But then again, I
ran into the EXACT same problem I had right before I applied the
unique_ptr<Expr> with std::variant<...> of expression types.
The main source of problem was the cyclic definition, in which the
`Expr` cannot be defined with a variant/union of an expression type
`BinaryExpr` of which its members were of type `Expr`.
This is basically resulted from C/C++ refusing to do double-parsing
during the tokenization process, which could then be easily avoiding
this sort of problem.

At this point, I'm almost calling it quit, and porting the whole
thing over to Odin. Even a simple act of using using unions and raw
pointers make the code immediately simpler and easier to read; no
visitor patterns, no move semantics on the expression pointers.
Things can just be a lot simpler than they actually are when it
comes to C++.
But maybe I'll checkout the C-based chapters of the book before
actually jump shipping. If I don't find the C solution interesting
enough, I AM moving everything to Odin.