Embed the source code line number info into the byte codes.
The possible line numbers are stored in a separate array from the
actual, individual byte codes, which has its own up sides and down
sides. In short, byte code storage is efficient, while the line
numbers storage is not. Could of a place of optimization down the
line, in order to store the line number with a "range" of byte code
instructions.