Binary operation support is done. Again, as simple as it goes.
Not even the disassembler needs any new functionality. Just
levaraging the all-constants-in-constant-pool mechanism to work out
the binary operations as well.
Except, the binary ops are not similar to the regular ASM binary
ops, where instead of doing `ADD <op1> <op2>`, the Lox binary
operation just pops the last 2 constants from the stack, do the op
and push the result back in. That's it.