Setup the basic structures of the compiler+lexer pipeline.
The basic structs of both Token and the Scanner are a bit different
from Lox, understandbly.
Both of them relies on referencing back to the original source line
(mainly cause C has no string types). So the source line has to live
longer than those of compiler, scanner/lexer instances.
The meat of the lexer grammer has not been put in yet.