Add identifier and keyword parsing in the Scanner/Lexer.
The 'keyword' identification mechanism is simply just walking down
a big, simple (yet not dumb) tree of possible characters in all the
keywords. Once each branch hits its node (and only when, not
partial) it'll bail out of the tree and mark it as a 'keyword'.
Otherwise, it's just an 'identifier'.

Like I said; simple, but not dumb.